Refactored for better code organization.
Used boltdb as an example.
