+++ request_test.go
@@ -1,0 +1,12 @@
+package ght_test
+
+import (
+	"net/http"
+	"reflect"
+	"regexp"
+	"testing"
+)
+
+
+func TestcheckRequest(t *testing.T) {
+}
